Moved files from src/kattis/truck/ to  src/kattis/truck/rec/ , added file for some heavy test
